Finished Candidate Generation
